function_source/footystats_get_league_list/requirements.txt
functions-framework
google-cloud-storage
requests